+++ chatGPT/Data/SwiftMarkdownRepository.swift
@@ -1,0 +1,55 @@
+import UIKit
+import Markdown
+
+final class SwiftMarkdownRepository: MarkdownRepository {
+    private let codeRegex = try! NSRegularExpression(pattern: "```(.*?)\\n([\\s\\S]*?)```", options: [])
+
+    func parse(_ markdown: String) -> NSAttributedString {
+        let matches = codeRegex.matches(in: markdown, options: [], range: NSRange(location: 0, length: markdown.utf16.count))
+        var parts: [NSAttributedString] = []
+        var currentLocation = markdown.startIndex
+
+        for match in matches {
+            guard let range = Range(match.range, in: markdown) else { continue }
+            let beforeText = String(markdown[currentLocation..<range.lowerBound])
+            if !beforeText.isEmpty {
+                parts.append(attributed(from: beforeText))
+            }
+
+            let codeRange = Range(match.range(at: 2), in: markdown)!
+            let code = String(markdown[codeRange])
+            let attachment = CodeBlockAttachment(code: code)
+            parts.append(NSAttributedString(attachment: attachment))
+
+            currentLocation = range.upperBound
+        }
+
+        let remaining = String(markdown[currentLocation...])
+        if !remaining.isEmpty {
+            parts.append(attributed(from: remaining))
+        }
+
+        let result = NSMutableAttributedString()
+        parts.forEach { result.append($0) }
+        while result.string.hasSuffix("\n") {
+            result.deleteCharacters(in: NSRange(location: result.length - 1, length: 1))
+        }
+        return result
+    }
+
+    private func attributed(from markdown: String) -> NSAttributedString {
+        var options = AttributedString.MarkdownParsingOptions()
+        options.interpretedSyntax = .inlineOnlyPreservingWhitespace
+        if let attr = try? AttributedString(markdown: markdown, options: options) {
+            let ns = NSMutableAttributedString(attr)
+            let range = NSRange(location: 0, length: ns.length)
+            ns.addAttribute(.font, value: UIFont.systemFont(ofSize: 16), range: range)
+            ns.addAttribute(.foregroundColor, value: UIColor.label, range: range)
+            return ns
+        }
+        return NSAttributedString(string: markdown, attributes: [
+            .font: UIFont.systemFont(ofSize: 16),
+            .foregroundColor: UIColor.label
+        ])
+    }
+}
